+++ TreningsApp2/app/src/main/java/com/example/kamil/treningsapp/Models/MealData.java
@@ -13,13 +13,13 @@ public class MealData {
     String name;
     String tag;
     int energy;
-    int protein;
-    int carbo;
-    int fat;
+    double protein;
+    double carbo;
+    double fat;
 
     public MealData(){}
 
-    public MealData(int iId, String day, String mealName, int weight, int foodId, String name, String tag, int energy, int protein, int carbo, int fat) {
+    public MealData(int iId, String day, String mealName, int weight, int foodId, String name, String tag, int energy, double protein, double carbo, double fat) {
         this.iId = iId;
         this.day = day;
         this.mealName = mealName;
@@ -32,7 +32,7 @@ public class MealData {
         this.carbo = carbo;
         this.fat = fat;
     }
-    public MealData(String day, String mealName, int weight, int foodId, String name, String tag, int energy, int protein, int carbo, int fat) {
+    public MealData(String day, String mealName, int weight, int foodId, String name, String tag, int energy, double protein, double carbo, double fat) {
         this.day = day;
         this.mealName = mealName;
         this.weight = weight;
@@ -109,27 +109,27 @@ public class MealData {
         this.energy = energy;
     }
 
-    public int getProtein() {
+    public double getProtein() {
         return protein;
     }
 
-    public void setProtein(int protein) {
+    public void setProtein(double protein) {
         this.protein = protein;
     }
 
-    public int getCarbo() {
+    public double getCarbo() {
         return carbo;
     }
 
-    public void setCarbo(int carbo) {
+    public void setCarbo(double carbo) {
         this.carbo = carbo;
     }
 
-    public int getFat() {
+    public double getFat() {
         return fat;
     }
 
-    public void setFat(int fat) {
+    public void setFat(double fat) {
         this.fat = fat;
     }
 }
